added changes to run on bsd

+++ config.h
@@ -137,12 +137,12 @@ const char *RESPONSE_403 = "403 - U better not go down this road!\r\n";
  * on new connection BUFFSIZE_READ -1 bytes are read from the socket,
  * everything extending that limit will be thrown away.
  */
-const uint32_t BUFFSIZE_READ = 2048;
+const size_t BUFFSIZE_READ = 2048;
 
 /*
  * if a file is transferd BUFFSIZE_WRITE describes the buffersize of
  * bytes read and then written to the socket.
  */
-const uint32_t BUFFSIZE_WRITE = 8192;
+const size_t BUFFSIZE_WRITE = 8192;
 
 #endif
